refactor: move compare_facts from model/ to compare/
Codex review on PR #1033: compare_facts()/FactComparison/FactComparability
decide whether an (old_fact, new_fact) pair differs -- exactly the "does
it differ" question model/AGENTS.md's own scoped contract reserves for
compare/, never model/. model/ owns what a fact is (Fact[T]/FactStatus),
not whether two of them disagree.

Moves the primitive to abicheck/compare/fact_comparison.py (compare/ may
depend on model/ per ADR-061 D1, which this module does for Fact/
FactStatus) and updates the two call sites (diff_types.py,
diff_param_qualifiers.py) plus its test suite (split into its own
tests/test_compare_fact_comparison.py, mirroring the source move) and the
plan doc/changelog references accordingly. No behavior change.
